// client/src/setup/store/actionsTypes/auth.actionsTypes.ts
export enum AuthActionTypes {
  REQUEST_SIGN = "REQUEST_SIGN",
  REQUEST_RIGISTER = "REQUEST_RIGISTER",
  REQUEST_SIGNOUT="REQUEST_SIGNOUT",
  REQUEST_GET_USERS = "REQUEST_GET_USERS",
  REQUEST_GET_USERS_PAGE = "REQUEST_GET_USERS_PAGE",
  CLEAR_SIGN_IN_FLAG = "CLEAR_SIGN_IN_FLAG",
  SUCCESS_SIGNOUT = "SUCCESS_SIGNOUT",
  CLEAR_USER_PAGE = "CLEAR_USER_PAGE",
  REQUEST_CONFIRMED = "REQUEST_CONFIRMED",
  FAILED_CONFIRMED = "FAILED_CONFIRMED",
  FAILED_RIGISTER = "FAILED_RIGISTER",
  SUCCESS_RIGISTER = "SUCCESS_RIGISTER",
  SUCCESS_CONFIRMED = "SUCCESS_CONFIRMED",
  SUCCESS_SIGN="SUCCESS_SIGN",
  FAILED_SIGN="FAILED_SIGN",
  FAILED_GET_USERS="FAILED_GET_USERS",
  FAILED_SIGNOUT="FAILED_SIGNOUT"
  
}