// client/src/setup/store/actionsTypes/companyProfile.actionsTypes.ts
export enum CompanyProfileType {
  REQUEST_GET_PROFILE_COMPANY = "REQUEST_GET_PROFILE_COMPANY ",
  SUCCESS_GET_PROFILE_COMPANY = "SUCCESS_GET_PROFILE_COMPANY",
  FAILED_GET_PROFILE_COMPANY = "FAILED_GET_PROFILE_COMPANY",
  REQUEST_UPDATE_PROFILE_COMPANY = "REQUEST_UPDATE_PROFILE_COMPANY ",
  SUCCESS_UPDATE_PROFILE_COMPANY = "SUCCESS_UPDATE_PROFILE_COMPANY",
  FAILED_UPDATE_PROFILE_COMPANY = "FAILED_UPDATE_PROFILE_COMPANY",
}