// client/src/setup/store/actionsTypes/userProfile.actionType.ts
export enum userProfileType {
    REQUEST_GET_PROFILE_USER = "REQUEST_GET_PROFILE_USER ",
    SUCCESS_GET_PROFILE_USER = "SUCCESS_GET_PROFILE_USER",
    FAILED_GET_PROFILE_USER = "FAILED_GET_PROFILE_USER",
    REQUEST_UPDATE_PROFILE_USER = "REQUEST_UPDATE_PROFILE_USER ",
    SUCCESS_UPDATE_PROFILE_USER = "SUCCESS_UPDATE_PROFILE_USER",
    FAILED_UPDATE_PROFILE_USER = "FAILED_UPDATE_PROFILE_USER",
  }
  